Update score type #1

+++ server/src/main/kotlin/com/coolcode/server/dto/ProgressDto.kt
@@ -1,10 +1,9 @@
 package com.coolcode.server.dto
 
-import java.math.BigDecimal
 import java.time.ZonedDateTime
 
 data class ProgressDto(
-    val score: BigDecimal,
+    val score: Int,
     var isPeerScoreOverridden: Boolean,
     val isMentorPasswordOverridden: Boolean,
     val isPeerAllScoresOverridden: Boolean,
